requirements.txt
# ===============================
# Core / Web
# ===============================
fastapi==0.110.3
uvicorn==0.29.0

# ===============================
# Data validation / config
# ===============================
pydantic==2.7.4
pydantic-settings==2.3.4

# ===============================
# LLM / Embeddings / LangChain ecosystem
# ===============================
langchain==0.2.14
langchain-community==0.2.12
langchain-groq==0.1.6
langchain-huggingface==0.0.3
langgraph==0.0.51
sentence-transformers==2.7.0
tiktoken==0.7.0

# ===============================
# Vector DB / Persistence / Clients
# ===============================
pinecone-client==3.2.2
supabase==2.4.3
redis==5.0.7
psycopg2-binary==2.9.9

# ===============================
# HTTP / I/O / Uploads
# ===============================
httpx==0.27.0
python-multipart==0.0.9
urllib3==2.2.2

# ===============================
# Auth / Security / Crypto
# ===============================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1

# ===============================
# Document processing / utils
# ===============================
pypdf==4.3.1
docx2txt==0.8

# ===============================
# Dev / Test
# ===============================
pytest==8.3.2



# fastapi
# uvicorn
# pydantic
# pydantic-settings
# langgraph
# langchain
# langchain-groq
# langchain-huggingface
# sentence-transformers
# supabase
# pinecone-client==3.2.2
# redis
# python-dotenv
# httpx
# python-multipart
# python-jose[cryptography]
# passlib[bcrypt]
# psycopg2-binary
# pytest
# # Added dependencies
# urllib3
# pypdf
# tiktoken
# langchain-community
# docx2txt
# email-validator
# pydantic[email]
